+++ sdk/ml/azure-ai-ml/azure/ai/ml/_schema/_distillation/__init__.py
@@ -5,7 +5,9 @@
 __path__ = __import__("pkgutil").extend_path(__path__, __name__)
 
 from .distillation_job import DistillationJobSchema
-from .distillation_types import EndpointRequestSettingsSchema, PromptSettingsSchema, TeacherModelSettingsSchema
+from .endpoint_request_settings import EndpointRequestSettingsSchema
+from .prompt_settings import PromptSettingsSchema
+from .teacher_model_settings import TeacherModelSettingsSchema
 
 __all__ = [
     "DistillationJobSchema",
