+++ sdk/ml/azure-ai-ml/azure/ai/ml/distillation/__init__.py
@@ -4,11 +4,9 @@
 
 __path__ = __import__("pkgutil").extend_path(__path__, __name__)
 
-from azure.ai.ml.entities._job.distillation.distillation_types import (
-    EndpointRequestSettings,
-    PromptSettings,
-    TeacherModelSettings,
-)
+from azure.ai.ml.entities._job.distillation.endpoint_request_settings import EndpointRequestSettings
+from azure.ai.ml.entities._job.distillation.prompt_settings import PromptSettings
+from azure.ai.ml.entities._job.distillation.teacher_model_settings import TeacherModelSettings
 
 from ._distillation import distillation
