+++ sdk/ml/azure-ai-ml/tests/distillation_job/unittests/test_distillation_job_schema.py
@@ -6,11 +6,9 @@ from azure.ai.ml.constants import DataGenerationTaskType, DataGenerationType
 from azure.ai.ml.entities import NoneCredentialConfiguration, ServerlessConnection, WorkspaceConnection
 from azure.ai.ml.entities._inputs_outputs import Input, Output
 from azure.ai.ml.entities._job.distillation.distillation_job import DistillationJob
-from azure.ai.ml.entities._job.distillation.distillation_types import (
-    EndpointRequestSettings,
-    PromptSettings,
-    TeacherModelSettings,
-)
+from azure.ai.ml.entities._job.distillation.endpoint_request_settings import EndpointRequestSettings
+from azure.ai.ml.entities._job.distillation.prompt_settings import PromptSettings
+from azure.ai.ml.entities._job.distillation.teacher_model_settings import TeacherModelSettings
 from azure.ai.ml.entities._job.finetuning.finetuning_job import FineTuningJob
 from azure.ai.ml.entities._job.resource_configuration import ResourceConfiguration
